Finished 4.21*: bloglist expansion, step9

+++ bloglist/controllers/blogs.js
@@ -38,7 +38,20 @@ blogsRouter.post("/", async (request, response) => {
 })
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(request.params.id)
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+  } else {
+    return response
+      .status(401)
+      .json({ error: "only the author can delete a blogpost" })
+  }
   response.status(204).end()
 })
